Extract existing handler to a service
